+++ usdSolidOcct/__init__.py
@@ -1,3 +1,4 @@
 from pxr import Tf, UsdSolid
+__version__ = "0.1.4"
 Tf.PreparePythonModule()
 del Tf, UsdSolid